+++ R/derive_merged.R
@@ -869,8 +869,8 @@ derive_var_merged_character <- function(dataset,
 
 #' Merge Lookup Table with Source Dataset
 #'
-#' Merge user-defined lookup table with the input dataset. Optionally output a
-#' list of observations from the input dataset that do not have corresponding
+#' Merge user-defined lookup table with the input dataset. Optionally print a
+#' list of records from the input dataset that do not have corresponding
 #' mapping from the lookup table.
 #'
 #' @param dataset_add Lookup table
@@ -885,13 +885,13 @@ derive_var_merged_character <- function(dataset,
 #' *Permitted Values*: TRUE, FALSE
 #'
 #' @inheritParams derive_vars_merged
-#'   ```
+#'
 #'
 #' @return The output dataset contains all observations and variables of the
-#' input dataset, and the variables specified in `new_vars`,
-#' mapped from the lookup table specified in `dataset_add`. Optionally prints
-#' a list of unique `by_vars` values that do not have corresponding records
-#' from the lookup table (by using option `print_not_mapped`).
+#' input dataset, and add the variables specified in `new_vars from the lookup
+#' table specified in `dataset_add`. Optionally prints a list of unique
+#' `by_vars` values that do not have corresponding records
+#' from the lookup table (by specifying `print_not_mapped = TRUE`).
 #'
 #' @author Annie Yang
 #'
@@ -903,27 +903,29 @@ derive_var_merged_character <- function(dataset,
 #' library(admiral.test)
 #' library(dplyr, warn.conflicts = FALSE)
 #' data("admiral_vs")
-#' vs <- convert_blanks_to_na(admiral_vs)
 #' param_lookup <- tibble::tribble(
-#'   ~VSTESTCD, ~PARAMCD, ~VSTEST, ~PARAMN,
-#'   "SYSBP", "SYSBP", "Systolic Blood Pressure", 1,
-#'   "WEIGHT", "WEIGHT", "Weight", 4,
-#'   "HEIGHT", "HEIGHT", "Height", 5,
-#'   "TEMP", "TEMP", "Temperature", 6,
-#'   "MAP", "MAP", "Mean Arterial Pressure", 7,
-#'   "BMI", "BMI", "Body Mass Index", 8,
-#'   "BSA", "BSA", "Body Surface Area", 9
+#'   ~VSTESTCD, ~VSTEST, ~PARAMCD, ~PARAM,
+#'   "SYSBP", "Systolic Blood Pressure", "SYSBP", "Systolic Blood Pressure (mmHg)",
+#'   "WEIGHT", "Weight", "WEIGHT", "Weight (kg)",
+#'   "HEIGHT", "Height", "HEIGHT", "Height (cm)",
+#'   "TEMP", "Temperature", "TEMP", "Temperature (C)",
+#'   "MAP", "Mean Arterial Pressure", "MAP", "Mean Arterial Pressure (mmHg)",
+#'   "BMI", "Body Mass Index", "BMI", "Body Mass Index(kg/m^2)",
+#'   "BSA", "Body Surface Area", "BSA", "Body Surface Area(m^2)"
 #' )
-#' derive_vars_merged_lookup(dataset = vs,
-#'                           dataset_add = param_lookup,
-#'                           by_vars = vars(VSTESTCD),
-#'                           new_vars = vars(PARAMCD),
-#'                           print_not_mapped = TRUE)
+#' derive_vars_merged_lookup(
+#'   dataset = admiral_vs,
+#'   dataset_add = param_lookup,
+#'   by_vars = vars(VSTESTCD),
+#'   new_vars = vars(PARAMCD),
+#'   print_not_mapped = TRUE
+#' )
 derive_vars_merged_lookup <- function(dataset,
                                       dataset_add,
                                       by_vars,
                                       new_vars = NULL,
-                                      print_not_mapped = TRUE){
+                                      print_not_mapped = TRUE) {
+  assert_logical_scalar(print_not_mapped)
   res <- derive_vars_merged(
     dataset,
     dataset_add,
@@ -933,13 +935,15 @@ derive_vars_merged_lookup <- function(dataset,
   )
 
   if (print_not_mapped) {
-    temp_not_mapped <- res %>% filter(is.na(temp_match_flag)) %>%
-      distinct(!!!by_vars) %>% unlist()
+    temp_not_mapped <- res %>%
+      filter(is.na(temp_match_flag)) %>%
+      distinct(!!!by_vars)
 
-    message("Note list of ", enumerate(vars2chr(by_vars)), " not mapped: ",
-            enumerate(temp_not_mapped, quote_fun = dquote))
+    message(
+      "List of ", enumerate(vars2chr(by_vars)), " not mapped: ", "\n",
+      paste0(capture.output(temp_not_mapped), collapse = "\n")
+    )
   }
 
-  res %>%select(-temp_match_flag)
-
+  res %>% select(-temp_match_flag)
 }